Changed double to float for deltaTime. Fixed params for particle and menu item draws. Passed in GameCore* to Screen_Base and all menu items.

+++ BitsAndPieces/ScoreChunk.cpp
@@ -78,7 +78,7 @@ void ScoreChunk::SetupWithDest( MyMeshText* pTextMesh, float x, float y, float d
     m_Justification = justification;
 }
 
-void ScoreChunk::Tick(double TimePassed)
+void ScoreChunk::Tick(float TimePassed)
 {
     float timepassed = (float)TimePassed;
     m_TimeAlive += timepassed;
